+++ core/constants.py
@@ -1,6 +1,7 @@
 SERVICE_NAME = 'great-cms'
 DASHBOARD_URL = '/dashboard/'
 LOGIN_URL = '/login/'
+EXPORT_PLAN_DASHBOARD_URL = '/export-plan/dashboard/'
 
 VIDEO_DURATION_DATA_ATTR_NAME = 'data-v-duration'
 
@@ -8,3 +9,6 @@ BACKLINK_QUERYSTRING_NAME = 'return-link'
 
 # Define a strict subset of rich-text features only includes linebreaks
 RICHTEXT_FEATURES__MINIMAL = ()
+
+LESSON_BLOCK = 'lesson'
+PLACEHOLDER_BLOCK = 'placeholder'