+++ tests/unit/core/test_middleware.py
@@ -7,7 +7,7 @@ from django.contrib.auth.models import AnonymousUser
 
 from core import helpers, middleware
 from tests.unit.core import factories
-from tests.unit.exportplan.factories import ExportPlanPageFactory, ExportPlanDashboardPageFactory
+from tests.unit.exportplan.factories import ExportPlanPageFactory, ExportPlanPseudoDashboardPageFactory
 from core.middleware import GADataMissingException
 
 
@@ -43,13 +43,15 @@ def test_stores_user_location_anon_user(mock_store_user_location, rf):
 
 
 @pytest.mark.django_db
-def test_user_specific_redirect_middleware(domestic_site, client):
+def test_user_specific_redirect_middleware(domestic_site, client, user, patch_export_plan):
     learn_page = factories.LandingPageFactory(parent=domestic_site.root_page, slug='learn')
     introduction_page = factories.ListPageFactory(
         parent=learn_page, slug='introduction', template='learn/automated_list_page.html'
     )
     categories_page = factories.CuratedListPageFactory(parent=learn_page, slug='categories')
     # Given the user has gone to /learn/introduction/
+
+    client.force_login(user)  # because unauthed users get redirected
     response = client.get(introduction_page.url)
 
     assert response.status_code == 200
@@ -68,7 +70,7 @@ def test_user_specific_redirect_exportplan_middleware_logged_in_company_name_set
     domestic_site, client, user, mock_get_company_profile, patch_export_plan
 ):
     exportplan_page = ExportPlanPageFactory(parent=domestic_site.root_page, slug='export-plan')
-    exportplan_dashboard_page = ExportPlanDashboardPageFactory(parent=exportplan_page, slug='dashboard')
+    exportplan_dashboard_page = ExportPlanPseudoDashboardPageFactory(parent=exportplan_page, slug='dashboard')
 
     # Given the user is logged in
     client.force_login(user)