+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### Fixed bugs
 - GP2-756 - Social logout and avatar fix
+- GP2-451 - Remove bottom margin from case study React component
+- GP2-451 - Remove empty step by step block on lesson page
+- GP2-357 - Heading padding below objective section
+- GP2-768 - but fix, dropdown no value
 - GP2-444 - select bug fix
 - GP2-761 - Added autoescape to render apostrophe for case study
 - GP2-754 - Pros and Cons layout tweaks again
@@ -68,9 +72,14 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+- GP2-735 - Update login/signup flow styling,
+- GP2-737 - Case Study enhancements in the CMS admin
+- GP2-593 - Dashboard tidy up
+- GP2-726 - Redirect unauthenticated users on CMS pages to the sign-in page
+- GP2-736 - Support lesson placeholders
 - GP2-443 - scroll to new route market
 - GP2-732 - marketing approach styling and content update
-- GP2-759 - Product search no-header 
+- GP2-759 - Product search no-header
 - GP2-759 - Product-search No results + feature flag to allow UR
 - GP2-755 - Add objectives section title
 - GP2-299 - Step by step reduce image size
@@ -284,7 +293,7 @@
 - GP2-709 - BE Country add required country signal
 - GP2-744 and GP2-744 - tighten up validation for CaseStudy.body StreamField
 - noticket - hide-next-link
-- GP2-731 -content-refresh
+- GP2-731 - content-refresh
 - GP2-781 - remove what's your business name prompt
 - GP2-699 - swamp route to market constants
 - GP2-743 - remove link on upload